refactor(array): move Array to its own namespace
